requirements.txt
boto3>=1.21.38,<1.22.0
python-dotenv>=0.20.0,<0.21.0
tweepy>=4.8.0,<4.9.0
